feat(AGOGOS-13): fix cli native build
It was found that the native build of the CLI has issues with installing
resources into the cluster because of the optimizations performed at the
build time by GraalVM which removes required classes from the resulting
binary making it unusable. Besides this -- even when we would have all
classes we would not be able to load arbitrary resources using tis
mechanism because we would not have the required model classes in the
classpath either case.

To overcome this issue we use the newly introduced in Fabric8 Kubernetes
Client 5.5 `GenericKubernetesResource`. This class can be used to
serialize/deserialize any Kubernetes objects. It does lack specific
handling for particular types of resources, but in our case we are not
really worried about it since the only thing we want to do is to apply
them into the cluster.

This commit uses `GenericKubernetesResource` as the transport for any
resources we will apply to the cluster.

But there is another problem. When we read resources from a file
we don't have the information about what is the endpoint in the
Kubernetes API so that we can create (or modify) the resource. To
determine this we use the Kubernetes OpenAPI data (available at
`/openapi/v2`). This provides indirect information what is the *plural*
of the resource type which is required to construct the URL.

Other changes:

* Introduction of `agogosctl load` command which can be used to load
  arbitrary content stored in YAML files.
* Cleanup of dependencies which made it possible to decrease the size of
  the native binary as well as speed up the startup and overcome an
  issue where two CLI processes could not be run in parallel.
* Quarkus upgrade to 2.0.0.Final
* Java Operator SDK update to 1.9.1.Final

https://issues.redhat.com/browse/AGOGOS-13