feat: init stage improvements and refactoring
This commits includes some cleanup in the code where
all of the clients for Agogos resources are replaced
with a single AgogosClient instance. It's much cleaner
and easier to understand how to interact with Agogos resources.

Teh init stage was updated to include kubectl binary so that it
can fetch the component definition itslef without passing
it as a parameter.

Installation was updated to provide a more reliable way to deploy
Knative Eventing Broker and added missing RoleBinding
for the agogos ServiceAccount.